pyproject.toml
[tool.poetry]
name = "flask-mongo-api-study"
version = "0.1.0"
description = ""
authors = ["luiz-zamboni <luiz.zamboni@test.com>"]
readme = "README.md"
packages = [{include="src"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
motor = "^3.6.0"
flask = {extras = ["async"], version = "^3.1.0"}
yml = "^0.0.1"
pyyaml = "^6.0.2"
uvicorn = "0.34.0"
asgiref = "3.8.1"
pydantic = "^2.10.4"
google-api-core = "2.25.1"
google-cloud-pubsub = "2.30.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
